Move the InitCommand class
The class has been moved into its own module.
